requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bert-score==0.3.13
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
datasets==2.13.1
dill==0.3.6
evaluate==0.4.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gin-config==0.5.0
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.0
joblib==1.3.1
kiwisolver==1.4.4
lmdb==1.4.0
matplotlib==3.7.2
methodtools==0.4.7
multidict==6.0.4
multiprocess==0.70.14
nltk==3.8.1
numpy==1.24.4
openai==0.27.8
packaging==23.1
pandas==2.0.3
path==16.6.0
path.py==12.5.0
Pillow==10.0.0
py-lz4framed==0.14.0
py-rouge==1.1
pyarrow==12.0.1
pyparsing==3.0.9
pyrouge==0.1.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rank-bm25==0.2.2
rapidfuzz==3.1.1
regex==2023.6.3
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wirerope==0.4.7
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
